Update Kestrel Overview to 6.0 (#24079)
* Update Kestrel Overview to 6.0

* .

* .

* .

* .
