--- aspnetcore/fundamentals/servers/kestrel/samples/6.x/Http3Sample/Http3Sample.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-</Project>
