--- aspnetcore/fundamentals/servers/kestrel/samples/6.x/Http3Sample/Program.cs
@@ -1,22 +1,0 @@
-using System;
-using System.Net;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Server.Kestrel.Core;
-
-#region snippet_UseHttp3
-var builder = WebApplication.CreateBuilder(args);
-builder.WebHost.ConfigureKestrel((context, options) =>
-{
-    options.ListenAnyIP(5001, listenOptions =>
-    {
-        // Use HTTP/3
-        listenOptions.Protocols = HttpProtocols.Http1AndHttp2AndHttp3;
-        listenOptions.UseHttps();
-    });
-});
-#endregion
-await using var app = builder.Build();
-app.MapGet("/", () => "Hello World!");
-await app.RunAsync();
