+++ aspnetcore/fundamentals/servers/kestrel/samples/6.x/KestrelSample/Program.cs
@@ -1,0 +1,8 @@
+// <snippet_CreateBuilder>
+var builder = WebApplication.CreateBuilder(args);
+var app = builder.Build();
+
+app.MapGet("/", () => "Hello World!");
+
+app.Run();
+// </snippet_CreateBuilder>
